backend/pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docflow-backend"
version = "0.0.0"
description = "DocFlow backend package"
authors = [{ name = "DocFlow Team" }]
dependencies = []

[project.optional-dependencies]
ci = [
    "sentry-sdk>=2.0",
    "prometheus-client>=0.20",
    "email-validator>=2.2",
    "pydantic-settings>=2.4",
    "opencv-python-headless>=4.10",
]

[tool.setuptools.packages.find]
include = ["backend"]

